refactor: 見積の子エンティティ構築を estimateChildBuilders へ抽出し EstimateFactory を委譲 facade 化
複製・新規生成と得意先改訂で二重化していた子構築（buildItem/buildSetGroups/
buildVariation）を domain/entities/estimateChildBuilders.ts のモジュールレベル純関数へ
一本化する土台を作る。EstimateFactory はこれらへ委譲する薄い facade にし、公開 API と
バレルの公開型は不変（挙動不変の純抽出）。

配置判断: 子エンティティは集約境界規約でバレル非公開のため、domain/entities/ 内からの
相対 import のみ許可される。EstimateFactory と同ディレクトリに置き Estimate は import
しない（estimateChildBuilders → 子エンティティの一方向・循環なし）。
