+++ src/server/subdomains/estimate/domain/entities/EstimateFactory.ts
@@ -18,12 +18,10 @@ import { Unit } from "../values/Unit";
 import { EstimateVariationCopy } from "../values/EstimateVariationCopy";
 import { EstimateVariationId } from "../values/EstimateVariationId";
 import { AfterRepairEstimateDetail } from "./AfterRepairEstimateDetail";
+import { buildVariation, buildVariationChildren } from "./estimateChildBuilders";
 import { Estimate } from "./Estimate";
-import { EstimateItem } from "./EstimateItem";
-import { EstimateSetGroup } from "./EstimateSetGroup";
 import { EstimateVariation, type TaxContext, type VariationContent } from "./EstimateVariation";
 import { RepairEstimateDetail } from "./RepairEstimateDetail";
-import { RevisedEstimateItemDetail } from "./RevisedEstimateItemDetail";
 
 /**
  * 集約外（アプリ層）から見積集約を生成するためのドメインファクトリ。
@@ -157,9 +155,7 @@ export class EstimateFactory {
       taxRoundingType: input.taxRoundingType,
     };
 
-    const variations = input.variations.map((variation) =>
-      EstimateFactory.buildVariation(variation, tax)
-    );
+    const variations = input.variations.map((variation) => buildVariation(variation, tax));
 
     return EstimateFactory.assembleEstimate(input, variations);
   }
@@ -182,7 +178,7 @@ export class EstimateFactory {
     };
 
     const built = input.variations.map((variation) => ({
-      variation: EstimateFactory.buildVariation(variation, tax),
+      variation: buildVariation(variation, tax),
       sourceVariationId: variation.sourceVariationId,
     }));
 
@@ -238,83 +234,13 @@ export class EstimateFactory {
    * 内容を組み立てるための入口（集約境界規約）。採番・差替えは集約ルートの責務。
    */
   static buildVariationContent(content: VariationContentDescriptor): VariationContent {
-    const normalItems = content.items.map((item) => EstimateFactory.buildItem(item));
-    const { groups, componentItems } = EstimateFactory.buildSetGroups(content.setGroups ?? []);
+    const { items, setGroups } = buildVariationChildren(content);
     return {
-      // 価格付き末端行（通常明細＋構成明細）を 1 配列に同居させる（ADR-0047）。
-      items: [...normalItems, ...componentItems],
-      setGroups: groups,
+      items,
+      setGroups,
       overallDiscount: content.overallDiscount,
       customerMemo: content.customerMemo,
       internalMemo: content.internalMemo,
     };
-  }
-
-  private static buildVariation(
-    variation: EstimateVariationDescriptor,
-    tax: TaxContext
-  ): EstimateVariation {
-    const normalItems = variation.items.map((item) => EstimateFactory.buildItem(item));
-    const { groups, componentItems } = EstimateFactory.buildSetGroups(variation.setGroups ?? []);
-    return EstimateVariation.create({
-      variationNumber: variation.variationNumber,
-      submissionType: variation.submissionType,
-      tax,
-      items: [...normalItems, ...componentItems],
-      setGroups: groups,
-      overallDiscount: variation.overallDiscount,
-      customerMemo: variation.customerMemo,
-      internalMemo: variation.internalMemo,
-    });
-  }
-
-  /**
-   * セット群記述子（入れ子の構成明細を持つ）から、群エンティティと平坦化した構成明細を構築する。
-   *
-   * **会員解決**: 構成明細を先に `EstimateItem.create` して id を確定させ、その id を群の
-   * `memberItemIds` へ配線する。群は実体ではなく順序付き id のみを保持し（案2-α）、構成明細の
-   * 実体は呼び出し側が `_items` へ同居させる。これにより `estimate_items` 1 表＋薄い衛星＋
-   * 所属交差表（ADR-0047）の形が、入れ子記述子から組み上がる。
-   */
-  private static buildSetGroups(descriptors: EstimateSetGroupDescriptor[]): {
-    groups: EstimateSetGroup[];
-    componentItems: EstimateItem[];
-  } {
-    const groups: EstimateSetGroup[] = [];
-    const componentItems: EstimateItem[] = [];
-    for (const descriptor of descriptors) {
-      const components = descriptor.components.map((c) => EstimateFactory.buildItem(c));
-      componentItems.push(...components);
-      groups.push(
-        EstimateSetGroup.create({
-          productId: descriptor.productId,
-          itemName: descriptor.itemName,
-          unit: descriptor.unit,
-          memberItemIds: components.map((c) => c.id),
-          customerMemo: descriptor.customerMemo,
-          internalMemo: descriptor.internalMemo,
-        })
-      );
-    }
-    return { groups, componentItems };
-  }
-
-  private static buildItem(item: EstimateItemDescriptor): EstimateItem {
-    return EstimateItem.create({
-      productId: item.productId,
-      sortOrder: item.sortOrder,
-      itemName: item.itemName,
-      quantity: item.quantity,
-      unit: item.unit,
-      unitPrice: item.unitPrice,
-      discountRate: item.discountRate,
-      itemDiscount: item.itemDiscount,
-      customerMemo: item.customerMemo,
-      internalMemo: item.internalMemo,
-      revisedDetail:
-        item.revisedDeliveryPrice != null
-          ? RevisedEstimateItemDetail.create(item.revisedDeliveryPrice)
-          : null,
-    });
   }
 }
